+++ changelog.md
@@ -2,6 +2,9 @@
   * Insta-blocking blocks more damage
 * Move creative event helper methods to InsaneLib
 
+## Upcoming
+* Attacking when blocking with crouch now puts the shield on a cooldown
+
 ## 2.1.0.1-beta
 * Fixed vanilla shield recipe not getting overridden